fix bug in Sampler-truncation; move TfSampler into models.tf
