+++ dataset/__init__.py
@@ -15,8 +15,7 @@ from .named_expr import B, C, F, L, V, R, W, P
 from .dsindex import DatasetIndex, FilesIndex
 from .decorators import action, inbatch_parallel, parallel, any_action_failed, mjit
 from .exceptions import SkipBatchException
-from .sampler import Sampler, ConstantSampler as CS
-from .sampler import NumpySampler as NS, TfSampler as TFS, HistoSampler as HS, ScipySampler as SS
+from .sampler import Sampler, ConstantSampler, NumpySampler, HistoSampler, ScipySampler
 
 
 __version__ = '0.3.0'
