+++ dataset/models/tf/__init__.py
@@ -19,3 +19,4 @@ from .densenet_fc import DenseNetFC, DenseNetFC56, DenseNetFC67, DenseNetFC103
 from .refinenet import RefineNet
 from .gcn import GlobalConvolutionNetwork as GCN
 from .encoder_decoder import EncoderDecoder
+from .tf_sampler import TfSampler
